+++ bin/bump-version.sh
@@ -4,6 +4,7 @@ OLD_VERSION="${1}"
 NEW_VERSION="${2}"
 
 export npm_config_git_tag_version=false
+npm --prefix packages/warden-service-api version "${NEW_VERSION}"
 npm --prefix packages/warden version "${NEW_VERSION}"
 
 pnpm install --frozen-lockfile